+++ frontend/src/services/customerService.js
@@ -1,0 +1,14 @@
+import { api } from './api';
+
+export const customerService = {
+  getCustomers: async (nameFilter = '') => {
+    const params = nameFilter ? { name: nameFilter } : {};
+    const response = await api.get('/customers', { params });
+    return response.data;
+  },
+
+  createCustomer: async (customerData) => {
+    const response = await api.post('/customers', customerData);
+    return response.data;
+  }
+};